+++ cmd/cmd.go
@@ -1,6 +1,6 @@
 /*
 	rulehunter - A server to find rules in data based on user specified goals
-	Copyright (C) 2016 vLife Systems Ltd <http://vlifesystems.com>
+	Copyright (C) 2016-2017 vLife Systems Ltd <http://vlifesystems.com>
 
 	This program is free software: you can redistribute it and/or modify
 	it under the terms of the GNU Affero General Public License as published by
@@ -16,54 +16,42 @@
 	along with this program; see the file COPYING.  If not, see
 	<http://www.gnu.org/licenses/>.
 */
-package main
+
+package cmd
 
 import (
-	"fmt"
+	"os"
+	"path/filepath"
+
 	"github.com/kardianos/service"
 	"github.com/vlifesystems/rulehunter/config"
 	"github.com/vlifesystems/rulehunter/html"
 	"github.com/vlifesystems/rulehunter/html/cmd"
 	"github.com/vlifesystems/rulehunter/logger"
+	"github.com/vlifesystems/rulehunter/program"
 	"github.com/vlifesystems/rulehunter/progress"
 	"github.com/vlifesystems/rulehunter/quitter"
-	"github.com/vlifesystems/rulehunter/watcher"
-	"log"
-	"os"
-	"path/filepath"
 )
 
-func main() {
-	flags := parseFlags(os.Args[1:])
-	l := logger.NewSvcLogger()
-	exitCode, err := subMain(flags, l)
-	if err != nil {
-		log.Fatal(err)
-	}
-	os.Exit(exitCode)
+type Setup struct {
+	prg *program.Program
+	svc service.Service
+	cfg *config.Config
 }
 
-// subMain is split from main to improve testing. quitter is used to stop
-// the routine.
-func subMain(
-	flags *cmdFlags,
+func InitSetup(
 	l logger.Logger,
-) (int, error) {
-	if err := handleFlags(flags); err != nil {
-		return 1, err
-	}
-
-	quit := quitter.New()
-	defer quit.Quit()
-
-	configFilename := filepath.Join(flags.configDir, "config.yaml")
+	q *quitter.Quitter,
+	configDir string,
+) (*Setup, error) {
+	configFilename := filepath.Join(configDir, "config.yaml")
 	config, err := config.Load(configFilename)
 	if err != nil {
-		return 1, errConfigLoad{filename: configFilename, err: err}
+		return nil, errConfigLoad{filename: configFilename, err: err}
 	}
 
 	if err := buildConfigDirs(config); err != nil {
-		return 1, err
+		return nil, err
 	}
 
 	htmlCmds := make(chan cmd.Cmd, 100)
@@ -72,74 +60,49 @@ func subMain(
 		htmlCmds,
 	)
 	if err != nil {
-		return 1, err
+		return nil, err
 	}
-	prg := newProgram(config, pm, l, quit)
+	prg := program.New(config, pm, l, q)
 
-	s, err := newService(prg, config, flags)
+	s, err := newService(prg, flagUser, configDir)
 	if err != nil {
-		return 1, err
+		return nil, err
 	}
 	svcLogger, err := s.Logger(nil)
 	if err != nil {
-		return 1, err
+		return nil, err
 	}
 
 	l.SetSvcLogger(svcLogger)
-	go l.Run(quit)
-	go html.Run(config, prg.progressMonitor, l, quit, htmlCmds)
-
-	if flags.install {
-		s.Uninstall()
-		if err = s.Install(); err != nil {
-			return 1, err
-		}
-	} else if flags.serve {
-		if err = s.Run(); err != nil {
-			return 1, err
-		}
-	} else {
-		if err := processDir(prg, config); err != nil {
-			return 1, fmt.Errorf("Errors while processing dir")
-		}
-	}
-	return 0, nil
-}
-
-func processDir(p *program, config *config.Config) error {
-	files, err := watcher.GetExperimentFiles(config.ExperimentsDir)
-	if err != nil {
-		return err
-	}
-	for _, file := range files {
-		if err := p.ProcessFile(file); err != nil {
-			return err
-		}
-	}
-	return nil
+	go l.Run(q)
+	go html.Run(config, pm, l, q, htmlCmds)
+	return &Setup{
+		prg: prg,
+		svc: s,
+		cfg: config,
+	}, nil
 }
 
 func newService(
-	prg *program,
-	cfg *config.Config,
-	flags *cmdFlags,
+	prg *program.Program,
+	user string,
+	configDir string,
 ) (service.Service, error) {
 	svcConfig := &service.Config{
 		Name:        "rulehunter",
 		DisplayName: "Rulehunter server",
-		Description: "Rulehunter finds rules in data based on user specified goals.",
+		Description: "Rulehunter finds rules in data based on user defined goals.",
 	}
 
-	if cfg.User != "" {
-		svcConfig.UserName = cfg.User
+	if user != "" {
+		svcConfig.UserName = user
 	}
 
-	if flags.configDir != "" {
-		svcConfig.Arguments =
-			[]string{fmt.Sprintf("-configdir=%s", flags.configDir), "-serve"}
+	svcConfig.Arguments = os.Args[1:]
+	if len(svcConfig.Arguments) >= 1 && svcConfig.Arguments[0] == "service" {
+		svcConfig.Arguments[0] = "serve"
 	}
-	s, err := service.New(prg, svcConfig)
-	return s, err
+	return service.New(prg, svcConfig)
 }
 
 func buildConfigDirs(cfg *config.Config) error {