+++ cmd/root_unix_test.go
@@ -1,6 +1,6 @@
 // +build darwin dragonfly freebsd linux nacl netbsd openbsd solaris
 
-package main
+package cmd
 
 import (
 	"os"
@@ -10,23 +10,16 @@ import (
 	"github.com/vlifesystems/rulehunter/internal/testhelpers"
 )
 
-func TestSubMain_interrupt(t *testing.T) {
+func TestRunRoot_interrupt(t *testing.T) {
 	cfgDir := testhelpers.BuildConfigDirs(t, false)
-	flags := &cmdFlags{install: false, serve: true, configDir: cfgDir}
 	defer os.RemoveAll(cfgDir)
-	mustWriteConfig(t, cfgDir, 100)
+	testhelpers.MustWriteConfig(t, cfgDir, 100)
 
 	l := testhelpers.NewLogger()
 	hasQuitC := make(chan bool)
 	go func() {
-		wantExitCode := 0
-		exitCode, err := subMain(flags, l)
-		if exitCode != wantExitCode {
-			t.Errorf("subMain(%v) exitCode: %d, want: %d",
-				flags, exitCode, wantExitCode)
-		}
-		if err != nil {
-			t.Errorf("subMain(%v): %s", flags, err)
+		if err := runRoot(l, cfgDir); err != nil {
+			t.Errorf("runRoot: %s", err)
 		}
 		hasQuitC <- true
 	}()
@@ -37,10 +30,9 @@ func TestSubMain_interrupt(t *testing.T) {
 		case <-interruptC:
 			interruptProcess(t)
 		case <-timeoutC:
-			t.Fatal("subMain() hasn't stopped")
+			t.Fatal("runRoot hasn't stopped")
 		case <-hasQuitC:
 			return
 		}
 	}
-
 }