+++ rulehunter_systemd_test.go
@@ -9,9 +9,9 @@ import "testing"
  *************************************/
 
 func startService(t *testing.T, serviceName string) {
-	runCmd(t, "systemctl", "start", serviceName)
+	runOSCmd(t, "systemctl", "start", serviceName)
 }
 
 func stopService(t *testing.T, serviceName string) {
-	runCmd(t, "systemctl", "stop", serviceName)
+	runOSCmd(t, "systemctl", "stop", serviceName)
 }