+++ rulehunter_upstart_test.go
@@ -9,9 +9,9 @@ import "testing"
  *************************************/
 
 func startService(t *testing.T, serviceName string) {
-	runCmd(t, "service", serviceName, "start")
+	runOSCmd(t, "service", serviceName, "start")
 }
 
 func stopService(t *testing.T, serviceName string) {
-	runCmd(t, "service", serviceName, "stop")
+	runOSCmd(t, "service", serviceName, "stop")
 }