+++ CHANGELOG.md
@@ -8,6 +8,7 @@
  * Up Go requirement to v1.8+
  * s/failure/error/ in activity
  * Add `category` to experiment files
+ * Change CLI to use a command structure. E.G. rulehunter serve --configdir=.
 
 ## 0.1 (7th July 2017)